Doc: Convert shadowsocks-libev man page to asciidoc
The port to asciidoc has some format change compared to old one.

The most obvious one is the options listed in SYNOPSIS.
Now the options list is not split into 2/3 columns, or we must use
asciidoc tables and introduce table frames.

Other small change includes the removal of AUTHOR sector, as it's not
an expendable method to update AUTHOR sector every time it gets updated.

Signed-off-by: Adam Anonymous <anonymous_temp_mail@yahoo.co.jp>